Make crate no_std compatible
`no_std` is used for constrained platforms, such as microcontrollers. As
TinyMT is a particularly small and efficient random number generator,
making it easily usable on such platforms is a good thing.

The changes required to make this crate compatible to `no_std` were
fairly small and do not affect `std` users:

- If a user wants to use the `std` features of the dependency `rand`,
  feature unification will cause `rand` to be built with the `std`
  feature without any issues.

- This crate (TinyMT) does not use any non-core features, so it is not
  impacted directly by `#![no_std]`.